Fresh eyes improvements!
Improves the instructions wording in the README, clarifying language
and emphasising the goal.

Adds some comments and spacing to the `main` method to direct the
player's attention to the function declaration that will lead them to
the code that needs to be changed.

Selecting discs / points in the visualisation shows the location of
that point alongside its label and (for a disc) radius.